# docs/swagger.yaml
basePath: /
definitions:
...
  types.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  types.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/types.APIError'
      success:
        type: boolean
    type: object
...
  types.FlowResponse:
    properties:
      call_data_hex:
        description: 调用数据
        type: string
      cancel_tx_hash:
        description: 取消交易哈希
        type: string
      cancelled_at:
        description: 取消时间
        type: string
      chain_id:
        description: 链ID
        type: integer
      contract_address:
        description: 合约地址
        type: string
      created_at:
        description: 创建时间
        type: string
      eta:
        description: 执行时间
        type: string
      execute_tx_hash:
        description: 执行交易哈希
        type: string
      executed_at:
        description: 执行时间
        type: string
      expired_at:
        description: 过期时间
        type: string
      flow_id:
        description: 流程ID
        type: string
      id:
        description: ID
        type: integer
      initiator_address:
        description: 发起者地址(FromAddress)
        type: string
      queue_tx_hash:
        description: 排队交易哈希
        type: string
      status:
        description: 状态
        type: string
      target_address:
        description: 目标地址
        type: string
      timelock_standard:
        description: Timelock标准
        type: string
      updated_at:
        description: 更新时间
        type: string
      value:
        description: 价值
        type: string
    type: object
...
  types.GetFlowListRequest:
    properties:
      page:
        description: 页码，默认为1
        type: integer
      page_size:
        description: 每页大小，默认为10，最大100
        type: integer
      standard:
        description: 标准compound, openzeppelin
        type: string
      status:
        description: 状态all, waiting, ready, executed, cancelled, expired
        type: string
    type: object
  types.GetFlowListResponse:
    properties:
      flows:
        description: 流程列表
        items:
          $ref: '#/definitions/types.FlowResponse'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  types.GetPublicSponsorsResponse:
    properties:
      partners:
        items:
          $ref: '#/definitions/types.SponsorInfo'
        type: array
      sponsors:
        items:
          $ref: '#/definitions/types.SponsorInfo'
        type: array
    type: object
...
  types.GetTransactionDetailRequest:
    properties:
      standard:
        description: 标准
        enum:
        - compound
        - openzeppelin
        type: string
      tx_hash:
        description: 交易哈希
        type: string
    required:
    - standard
    - tx_hash
    type: object
  types.GetTransactionDetailResponse:
    properties:
      detail:
        allOf:
        - $ref: '#/definitions/types.TimelockTransactionDetail'
        description: 交易详情
    type: object
...
  types.SponsorInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      link:
        type: string
      logo_url:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      type:
        $ref: '#/definitions/types.SponsorType'
    type: object
  types.SponsorType:
    enum:
    - partner
    - sponsor
    type: string
    x-enum-comments:
      SponsorTypePartner: 生态伙伴
      SponsorTypeSponsor: 赞助方
    x-enum-varnames:
    - SponsorTypePartner
    - SponsorTypeSponsor
...
  types.TimelockTransactionDetail:
    properties:
      block_number:
        description: 区块高度
        type: integer
      block_timestamp:
        description: 区块时间
        type: string
      chain_id:
        description: 链ID
        type: integer
      chain_name:
        description: 链名称
        type: string
      contract_address:
        description: 合约地址
        type: string
      from_address:
        description: 发起地址
        type: string
      to_address:
        description: 接收地址
        type: string
      tx_hash:
        description: 交易哈希
        type: string
      tx_status:
        description: 交易状态（success, failed）
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: TimeLocker Backend API
  title: TimeLocker Backend API
  version: "1.0"
paths:
...
  /api/v1/flows/list:
    post:
      consumes:
      - application/json
      description: 获取与用户相关的timelock流程列表，包括发起的和有权限管理的
      parameters:
      - description: 查询参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/types.GetFlowListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.GetFlowListResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取与用户相关的流程列表
      tags:
      - Flow
  /api/v1/flows/transaction/detail:
    post:
      consumes:
      - application/json
      description: 根据交易哈希和标准获取交易详情。standard 仅支持 compound/openzeppelin；tx_hash 必须为
        0x 开头的64位十六进制。
      parameters:
      - description: 请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.GetTransactionDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.GetTransactionDetailResponse'
              type: object
        "400":
          description: 请求参数错误（INVALID_STANDARD / INVALID_TX_HASH）
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: 交易不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 获取交易详情
      tags:
      - Flow
  /api/v1/sponsors/public:
    post:
      consumes:
      - application/json
      description: 获取所有激活的赞助方和生态伙伴信息，用于在前端展示。返回的数据按照排序权重和创建时间排序。
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取赞助方和生态伙伴列表
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.GetPublicSponsorsResponse'
              type: object
        "500":
          description: 获取赞助方列表失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 获取公开的赞助方和生态伙伴列表
      tags:
      - Sponsors
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"